Refactor to have the page numbers only with EPUB
